+++ cdp-agentkit-core/typescript/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 ## Unreleased
 
+## [0.0.12] - 2025-01-17
+
 ### Added
 
 - Added `get_balance_nft` action.
 - Added `transfer_nft` action.
 - Added `pyth_fetch_price_feed_id` action to fetch the price feed ID for a given token symbol from Pyth.
 - Added `pyth_fetch_price` action to fetch the price of a given price feed from Pyth.
+- Bump dependency `coinbase-sdk-nodejs` to version `0.14.1`.
+
+### Fixed
+
+- Allow wallet mnemonic seed import to optionally accept `networkId` input argument.
 
 ## [0.0.11] - 2025-01-13