+++ twitter-langchain/python/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## [0.0.9] - 2025-01-17
+
+### Added
+
+- Bump dependency `cdp-agentkit-core` to version `0.0.9`.
+
 ## [0.0.8] - 2025-01-13
 
 ### Added